Working

+++ database/migrations/2020_04_13_124843_create_love_table.php
@@ -17,7 +17,7 @@ class CreateLoveTable extends Migration
             $table->id();
             $table->string("name_1");
             $table->string("name_2");
-            $table->double('perc', 3, 2)->default('100');;
+            $table->integer('perc')->default('100');;
             $table->timestamps();
         });
     }
